+++ lumenairy/propagators/__init__.py
@@ -50,7 +50,9 @@ from .ensemble import propagate_ensemble
 from .gbd import (
     BeamletBundle,
     apply_abcd_to_beamlets,
+    apply_aperture_to_beamlets,
     apply_thin_lens_to_beamlets,
+    decompose_field_adaptive,
     decompose_field_to_beamlets,
     propagate_beamlets_freespace,
     propagate_gbd,
@@ -108,6 +110,8 @@ __all__ = [
     'PropagationResult',
     'BeamletBundle',
     'decompose_field_to_beamlets',
+    'decompose_field_adaptive',
+    'apply_aperture_to_beamlets',
     'propagate_beamlets_freespace',
     'apply_thin_lens_to_beamlets',
     'reconstruct_field_from_beamlets',